+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add constraints for dependencies in setup.py
+- Cache request to the API component registry
 
 ## [1.1.1] 2022-05-31